apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
redis==5.2.1
sqlalchemy==2.0.36
alembic==1.14.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery[redis]==5.4.0
httpx==0.28.1
stripe==8.4.0
feedparser==6.0.11
ruff==0.8.6
black==24.10.0
pytest==8.3.4
pytest-asyncio==0.25.2
